chanage the structure of the authOptions to match the condition when deploy
